app/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
click==8.1.7
exceptiongroup==1.1.3
fastapi==0.104.0
greenlet==3.0.0
h11==0.14.0
idna==3.4
pika==1.3.2
psycopg2==2.9.9
psycopg2-binary==2.9.9
pydantic==2.4.2
pydantic_core==2.10.1
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
python-dotenv==1.0.0
sniffio==1.3.0
SQLAlchemy==2.0.22
starlette==0.27.0
typing_extensions==4.8.0
uvicorn==0.23.2
